Improve memory profiling
Produce a profile right after loading a trace, so that the trace data is
actually still live, in addition to the trace we emit when terminating,
which captures all allocations that have been made.

Also make sure to run GC before writing the heap profile, to write up to
date data.