Don't track events per span
Instead of tracking events per span, track them per goroutine and use
binary search to find the events relevant to a given span. This greatly
reduces the number of slices we have to allocate, as there are much
fewer goroutines than there are spans. It also reduces the size of the
Span struct by 24 bytes, which can account for gigabytes of data in
large traces.