trace: represent P using int32 instead of uint32
Not all events have a P associated with them. For example, when
runtime.sysmon – which runs without a P – wakes up the scavenger, it
unblocks a number of goroutines.

The tracer emits those events with a pid of -1 (encoded as a uint64). We
use an int32 so that we can check for -1 instead of having to think
about MaxInt32. The original code used int, before we (somewhat
incorrectly) changed it to uint32.