trace: read file into memory
Loading the whole file into memory simplifies our code and improves
performance. Instead of the comparatively complex logic in bufio.Reader,
we only have to track an offset and a length now. It also allows all
functions concerned with reading bytes to be inlined.

Parsing a trace already needs roughly 20x the size of the input as
memory to represent all events. Temporarily needing 21x instead, to keep
the file in memory, seems negligible. Whether a 300 MB file needs 6 GB
or 6.3 GB to parse makes little difference to the user, making this a
fair CPU vs memory tradeoff.