Implement panning by holding the spacebar
This isn't working perfectly yet because Gio doesn't let us get a
pointer grab in response to key presses, which means the window can lose
focus while the spacebar is pressed. This is very annoying with focus
follows mouse on Linux, but probably less of a problem for most macOS
users, who need this function the most, because they lack a middle mouse
button.

Updates: gh-11