Improve input scheme for touchpad and macOS users
Most of our testers have complained about the use of the middle mouse
button, either because they were using touchpads, or because their
Apple-specific mice don't have one. Therefore we're switching to a
scheme focused on the left mouse button, which is now responsible for
panning, zooming, and clicking things.

As part of that change we've switched most code over to using
gesture.Click and gesture.Drag. This way, dragging and clicking things
don't interfere.

Because touchpad users are accustomed to "flinging" to scroll, we've
changed scrolling to actually scroll the timeline. Zooming now requires
Ctrl+Scroll. I am hoping that this magically makes flinging work on
macOS, but I couldn't test it.

Finally, all places that expected the Ctrl modifier now also accept the
Command modifier.