Don't use map for span PCs
Instead of tracking span PCs using a map, add MetadataSelector, which
works similarly to SpanSelector. This allows us to store span metadata
in slices.

One downside of this approach is that the metadata really belongs to the
container, such as a Track. We can't pass a single span to a function
and have it operate on the metadata; instead we need to pass a
MetadataSelector. On top of that, this approach only permits one type of
metadata per container.