Don't do any layout or shaping work concurrently with the window's goroutine
All uses of the theme.Window should happen as part of the window's
layout, not concurrently with it. In particular, futures shouldn't be
shaping text.

A text shaper isn't safe for concurrent use. Also, doing layout outside
of the actual Layout function tree encourages caching layout
permanently, without the use of a proper cache key. It also interacts
poorly with widgets that migrate between windows.