trace: add missing links for Waiting, InSyscall events
Goroutine events are normally linked so that one can follow the sequence
of state changes for a goroutine. However, there are a few links
missing:

Currently GoCreate events are only linked to the next GoStart. For
goroutines blocked at the beginning of the trace, the next event will be
a GoWaiting/GoInSyscall rather than a GoStart. This commit adds a link
that was missing in that case.

This commit also adds a link from GoInSyscall to the next GoSysExit,
matching what is done for GoSysCall events.