Cache merged span colors using interval tree
Contiguous spans can be represented as intervals of the first start
timestamp and last end timestamp. If an interval represents a merged
span with multiple colors, then any interval that is a superset of it
must also be a merged span with multiple colors. This allows us to cache
fewer overall entries while simultaneously having more cache hits.