Redesign spans and events abstraction
Instead of using ptrace.Spans in Gotraceui, introduce a new abstraction
that knows about and stores the timeline and track that items belong to.
This allows us to pass spans and events around, without having to
manually pass around the containing timelines and tracks. This also
increases the power of merging and filtering, as those can now deal with
multiple containers.