Compress stack tracks
For each goroutine timeline, there can be up to 64¹ tracks representing
the stack traces of events. As most events have stack traces, this can
be a 64x increase in memory used for goroutine tracks.

To avoid this size increase, we compress the spans needed for these
tracks, using a combination of delta encoding, zigzag encoding, and
simple-8b encoding for numeric data, and simple bitpacking for booleans.

When displaying the tracks and thus accessing the spans, they are
decompressed and cached until the tracks are no longer visible.

This marginally slows down loading of traces, but has negligible effect
on the UI's performance. On a Ryzen 3950x we are able to decode over a
billion 64-bit integers per second per core.

We considered compressing all tracks, but the data for "normal" tracks
currently comes from the ptrace package and is part of the result of
ptrace.Parse. In the future we might choose to compress the original
data and then discard it.