Ensure growing slices zeroes memory correctly
slices.Grow is only concerned with growing the capacity, assuming that
the new capacity will be appeneded over. For our use-case, which is to
increase the length of a slice, we'd need to zero the newly made
available memory. Instead of working around slices.Grow we introduce our
own helpers.