Implement custom compression for textures
Switch to Snappy for encoding data. Snappy is 1-2 orders of magnitude
faster than DEFLATE while producing less garbage. For favourable inputs,
Snappy's compression ratio is 50-100% that of DEFLATE. For unfavourable
inputs, it can be as bad as 12%.

To counteract Snappy's weakness, we notice that a lot of our textures
begin or end with a long stretch of a single color. We detect these
prefixes and suffixes and compress them down to a single color and a
repetition count. This is a very limited form of RLE. The remainder of
the data is compressed with Snappy, or stored uncompressed if the
compressed data is larger than the uncompressed data.

For favourable inputs with long prefixes and/or suffixes, this approach
compresses an order of magnitude faster than Snappy, while producing
compression ratios comparable to or significantly exceeding DEFLATE, by
up to 700%.

For unfavourable inputs, our approach has results identical to using
just Snappy.

On average over all textures, our approach has nearly identical
compression ratios to DEFLATE, while being 2 orders of magnitude faster
to compress and an order of magnitude faster to decompress while
producing significantly less garbage.