Rely entirely on textures for drawing spans
There's little point in drawing spans as rectangles on top of the
textures as they will look the same.

The only issue are rounding errors / conflation artifacts. We address
this by assuming at most two textures, and nudging the second one
slightly to the left. Not a pretty solution, but it works.