Don't copy spans
While it is unfortunate to pass pointers to a pure data type like
ptrace.Span, all of the copying of spans adds up and is noticeable in
CPU profiles.

Copying spans is especially bad because Go rarely optimizes the copy
away. For example, for 'spans.At(0).Start', the caller still copies the
entire span out of the return value, even though we only need one field
of it.