Always draw span borders, even if they occupy the entire span's width
Not having borders around some spans looks inconsistent and like they're
not a proper part of the timeline. We'd rather cover up the handful of
colored pixels with borders than not have borders.

This also fixes a bug when a borderless span was left to a bordered
span, where the right span was missing its left border, thinking the
span to its left had a right border.