Compute stack tracks on the fly
Storing all stack tracks, even in a compressed format, uses significant
amounts of memory. When we added support for compression, we also added
support for async creation of spans. Instead of storing compressed
stacks, we can just compute them on demand, and throw them away when the
tracks go out of view. In most cases, this is fast enough to finish
within a single frame, anyway. This saves up to 30% of memory.