Reimplement plotting
The core change is that we no longer use the Visvalingam-Whyatt
algorithm but instead pick up to four values for each pixel (bucket):
the first value, the minimum, the maximum, and the last value. This is
inspired by M4[1] and ensures a near perfect rendering of the plot (with
minor differences in anti-aliasing compared to rendering all points).

Because the new downsampling is much faster to compute, we only cache
one downsampling, 32k pixels worth of the whole plot. This can be used
when viewing the whole plot or when zoomed in slightly. This handles the
case where otherwise we would have to downsample millions of points
dynamically. We downsample the cached downsampling again when rendering
it, down to the actual number of visible pixels.

The twice-downsampled data isn't perfectly aligned with the pixel grid and
thus doesn't meet the requirements of M4, but for our interactive
purposes, the rendering looks very good and stable.

Downsampling using (first, min, max, last) depends on the fact that we
pan by whole multiples of nsPerPx, as this avoids points alternating
between buckets as we pan, which would lead to a shimmering effect.

Due to oddities in Gio's renderer and surprising performance cliffs, we
still remove "insignificant" points before rendering them. This is
generally visually imperceptible, but sometimes it introduces very minor
differences. We hope to eliminate this step eventually.

We've changed the representation of metrics from AoS to SoA to improve
the performance of argminmax and to prepare for a possible SIMD
implementation.

We no longer simulate filled lines using rectangles (which never
correctly worked for non-stairstep rendering) but use the 'curve'
package to compute accurate stroked lines. We do not use Gio's built-in
stroking because it is less performant, less accurate, and has run into
infinite loops that cause OOMs.

Combined, these changes massively reduce memory usage and load times for
large traces and produce more stable plots with fewer artifacts.

[1] https://doi.org/10.14778/2732951.2732953