Use a fork of Gio 0.4.1
Newer versions of Gio make significant changes to their API and we don't
want to invest time in porting our code to those, as we're expecting to
switch to an entirely library "soon." However, Gio 0.4.1 doesn't build
with Go 1.24 due to methods being declared on aliases to C types. This
issue is fixed in newer versions and we backport the fixes to a fork of
Gio 0.4.1.

Closes: gh-178